Crate aws_sdk_workmail

Amazon WorkMail is a secure, managed business email and calendaring service with support for existing desktop and mobile email clients. You can access your email, contacts, and calendars using Microsoft Outlook, your browser, or other native iOS and Android email applications. You can integrate WorkMail with your existing corporate directory and control both the keys that encrypt your data and the location in which your data is stored.

The WorkMail API is designed for the following scenarios:

Listing and describing organizations

Managing users

Managing groups

Managing resources

All WorkMail API operations are Amazon-authenticated and certificate-signed. They not only require the use of the AWS SDK, but also allow for the exclusive use of AWS Identity and Access Management users and roles to help facilitate access, trust, and permission policies. By creating a role and allowing an IAM user to access the WorkMail site, the IAM user gains full administrative visibility into the entire WorkMail organization (or as set in the IAM policy). This includes, but is not limited to, the ability to create, update, and delete users, groups, and resources. This allows developers to perform the scenarios listed above, as well as give users the ability to grant access on a selective basis using the IAM model.

Re-exports

pub use config::Config;

pub use client::Client;

Modules

client

config

error

input

model

operation

output

Structs

Blob

ByteStream

Stream of binary data

Credentials

AWS SDK Credentials

Endpoint

API Endpoint

Region

The region to send requests to.

Enums

Error

SdkError

Failed SDK Result

Statics

PKG_VERSION