Module aws_sdk_workmail::error

Modules

directory_in_use_exception

See DirectoryInUseException

directory_service_authentication_failed_exception

See DirectoryServiceAuthenticationFailedException

directory_unavailable_exception

See DirectoryUnavailableException

email_address_in_use_exception

See EmailAddressInUseException

entity_already_registered_exception

See EntityAlreadyRegisteredException

entity_not_found_exception

See EntityNotFoundException

entity_state_exception

See EntityStateException

invalid_configuration_exception

See InvalidConfigurationException

invalid_parameter_exception

See InvalidParameterException

invalid_password_exception

See InvalidPasswordException

limit_exceeded_exception

See LimitExceededException

mail_domain_not_found_exception

See MailDomainNotFoundException

mail_domain_state_exception

See MailDomainStateException

name_availability_exception

See NameAvailabilityException

organization_not_found_exception

See OrganizationNotFoundException

organization_state_exception

See OrganizationStateException

reserved_name_exception

See ReservedNameException

resource_not_found_exception

See ResourceNotFoundException

too_many_tags_exception

See TooManyTagsException

unsupported_operation_exception

See UnsupportedOperationException

Structs

AssociateDelegateToResourceError

AssociateMemberToGroupError

CancelMailboxExportJobError

CreateAliasError

CreateGroupError

CreateMobileDeviceAccessRuleError

CreateOrganizationError

CreateResourceError

CreateUserError

DeleteAccessControlRuleError

DeleteAliasError

DeleteGroupError

DeleteMailboxPermissionsError

DeleteMobileDeviceAccessRuleError

DeleteOrganizationError

DeleteResourceError

DeleteRetentionPolicyError

DeleteUserError

DeregisterFromWorkMailError

DescribeGroupError

DescribeMailboxExportJobError

DescribeOrganizationError

DescribeResourceError

DescribeUserError

DirectoryInUseException

The directory is already in use by another WorkMail organization in the same account and Region.

DirectoryServiceAuthenticationFailedException

The directory service doesn't recognize the credentials supplied by WorkMail.

DirectoryUnavailableException

The directory is unavailable. It might be located in another Region or deleted.

DisassociateDelegateFromResourceError

DisassociateMemberFromGroupError

EmailAddressInUseException

The email address that you're trying to assign is already created for a different

EntityAlreadyRegisteredException

The user, group, or resource that you're trying to register is already

EntityNotFoundException

The identifier supplied for the user, group, or resource does not exist in your

EntityStateException

You are performing an operation on a user, group, or resource that isn't in the

GetAccessControlEffectError

GetDefaultRetentionPolicyError

GetMailboxDetailsError

GetMobileDeviceAccessEffectError

InvalidConfigurationException

The configuration for a resource isn't valid. A resource must either be able to

InvalidParameterException

One or more of the input parameters don't match the service's restrictions.

InvalidPasswordException

The supplied password doesn't match the minimum security constraints, such as length

LimitExceededException

The request exceeds the limit of the resource.

ListAccessControlRulesError

ListAliasesError

ListGroupMembersError

ListGroupsError

ListMailboxExportJobsError

ListMailboxPermissionsError

ListMobileDeviceAccessRulesError

ListOrganizationsError

ListResourceDelegatesError

ListResourcesError

ListTagsForResourceError

ListUsersError

MailDomainNotFoundException

For an email or alias to be created in Amazon WorkMail, the included domain must be defined

MailDomainStateException

After a domain has been added to the organization, it must be verified. The domain is

NameAvailabilityException

The user, group, or resource name isn't unique in Amazon WorkMail.

OrganizationNotFoundException

An operation received a valid organization identifier that either doesn't belong or

OrganizationStateException

The organization must have a valid state to perform certain

PutAccessControlRuleError

PutMailboxPermissionsError

PutRetentionPolicyError

RegisterToWorkMailError

ReservedNameException

This user, group, or resource name is not allowed in Amazon WorkMail.

ResetPasswordError

ResourceNotFoundException

The resource cannot be found.

StartMailboxExportJobError

TagResourceError

TooManyTagsException

The resource can have up to 50 user-applied tags.

UnsupportedOperationException

You can't perform a write operation against a read-only directory.

UntagResourceError

UpdateMailboxQuotaError

UpdateMobileDeviceAccessRuleError

UpdatePrimaryEmailAddressError

UpdateResourceError

Enums

AssociateDelegateToResourceErrorKind

AssociateMemberToGroupErrorKind

CancelMailboxExportJobErrorKind

CreateAliasErrorKind

CreateGroupErrorKind

CreateMobileDeviceAccessRuleErrorKind

CreateOrganizationErrorKind

CreateResourceErrorKind

CreateUserErrorKind

DeleteAccessControlRuleErrorKind

DeleteAliasErrorKind

DeleteGroupErrorKind

DeleteMailboxPermissionsErrorKind

DeleteMobileDeviceAccessRuleErrorKind

DeleteOrganizationErrorKind

DeleteResourceErrorKind

DeleteRetentionPolicyErrorKind

DeleteUserErrorKind

DeregisterFromWorkMailErrorKind

DescribeGroupErrorKind

DescribeMailboxExportJobErrorKind

DescribeOrganizationErrorKind

DescribeResourceErrorKind

DescribeUserErrorKind

DisassociateDelegateFromResourceErrorKind

DisassociateMemberFromGroupErrorKind

GetAccessControlEffectErrorKind

GetDefaultRetentionPolicyErrorKind

GetMailboxDetailsErrorKind

GetMobileDeviceAccessEffectErrorKind

ListAccessControlRulesErrorKind

ListAliasesErrorKind

ListGroupMembersErrorKind

ListGroupsErrorKind

ListMailboxExportJobsErrorKind

ListMailboxPermissionsErrorKind

ListMobileDeviceAccessRulesErrorKind

ListOrganizationsErrorKind

ListResourceDelegatesErrorKind

ListResourcesErrorKind

ListTagsForResourceErrorKind

ListUsersErrorKind

PutAccessControlRuleErrorKind

PutMailboxPermissionsErrorKind

PutRetentionPolicyErrorKind

RegisterToWorkMailErrorKind

ResetPasswordErrorKind

StartMailboxExportJobErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateMailboxQuotaErrorKind

UpdateMobileDeviceAccessRuleErrorKind

UpdatePrimaryEmailAddressErrorKind

UpdateResourceErrorKind