Module aws_sdk_workmail::model

Modules

access_control_rule

See AccessControlRule

booking_options

See BookingOptions

delegate

See Delegate

domain

See Domain

folder_configuration

See FolderConfiguration

group

See Group

mailbox_export_job

See MailboxExportJob

member

See Member

mobile_device_access_matched_rule

See MobileDeviceAccessMatchedRule

mobile_device_access_rule

See MobileDeviceAccessRule

organization_summary

See OrganizationSummary

permission

See Permission

resource

See Resource

tag

See Tag

user

See User

Structs

AccessControlRule

A rule that controls access to an Amazon WorkMail organization.

BookingOptions

At least one delegate must be associated to the resource to disable automatic replies

Delegate

The name of the attribute, which is one of the values defined in the UserAttribute

Domain

The domain to associate with an Amazon WorkMail organization.

FolderConfiguration

The configuration applied to an organization's folders by its retention

Group

The representation of an Amazon WorkMail group.

MailboxExportJob

The details of a mailbox export job, including the user or resource ID associated

Member

The representation of a user or group.

MobileDeviceAccessMatchedRule

The rule that a simulated user matches.

MobileDeviceAccessRule

A rule that controls access to mobile devices for an Amazon WorkMail group.

OrganizationSummary

The representation of an organization.

Permission

Permission granted to a user, group, or resource to access a certain aspect of

Resource

The representation of a resource.

Tag

Describes a tag applied to a resource.

User

The representation of an Amazon WorkMail user.

Enums

AccessControlRuleEffect

EntityState

FolderName

MailboxExportJobState

MemberType

MobileDeviceAccessRuleEffect

PermissionType

ResourceType

RetentionAction

UserRole