Module aws_sdk_workmail::operation

Structs

AssociateDelegateToResource

Adds a member (user or group) to the resource's set of delegates.

AssociateMemberToGroup

Adds a member (user or group) to the group's set.

CancelMailboxExportJob

Cancels a mailbox export job.

CreateAlias

Adds an alias to the set of a given member (user or group) of Amazon WorkMail.

CreateGroup

Creates a group that can be used in Amazon WorkMail by calling the RegisterToWorkMail operation.

CreateMobileDeviceAccessRule

Creates a new mobile device access rule for the specified Amazon WorkMail organization.

CreateOrganization

Creates a new Amazon WorkMail organization. Optionally, you can choose to associate an existing AWS Directory Service directory with your organization. If an AWS Directory Service directory ID is specified, the organization alias must match the directory alias. If you choose not to associate an existing directory with your organization, then we create a new Amazon WorkMail directory for you. For more information, see Adding an organization in the Amazon WorkMail Administrator Guide.

CreateResource

Creates a new Amazon WorkMail resource.

CreateUser

Creates a user who can be used in Amazon WorkMail by calling the RegisterToWorkMail operation.

DeleteAccessControlRule

Deletes an access control rule for the specified WorkMail organization.

DeleteAlias

Remove one or more specified aliases from a set of aliases for a given

DeleteGroup

Deletes a group from Amazon WorkMail.

DeleteMailboxPermissions

Deletes permissions granted to a member (user or group).

DeleteMobileDeviceAccessRule

Deletes a mobile device access rule for the specified Amazon WorkMail organization.

DeleteOrganization

Deletes an Amazon WorkMail organization and all underlying AWS resources managed by Amazon WorkMail as part of the organization. You can choose whether to delete the associated directory. For more information, see Removing an organization in the Amazon WorkMail Administrator Guide.

DeleteResource

Deletes the specified resource.

DeleteRetentionPolicy

Deletes the specified retention policy from the specified organization.

DeleteUser

Deletes a user from Amazon WorkMail and all subsequent systems. Before you can delete a

DeregisterFromWorkMail

Mark a user, group, or resource as no longer used in Amazon WorkMail. This action

DescribeGroup

Returns the data available for the group.

DescribeMailboxExportJob

Describes the current status of a mailbox export job.

DescribeOrganization

Provides more information regarding a given organization based on its

DescribeResource

Returns the data available for the resource.

DescribeUser

Provides information regarding the user.

DisassociateDelegateFromResource

Removes a member from the resource's set of delegates.

DisassociateMemberFromGroup

Removes a member from a group.

GetAccessControlEffect

Gets the effects of an organization's access control rules as they apply to a

GetDefaultRetentionPolicy

Gets the default retention policy details for the specified organization.

GetMailboxDetails

Requests a user's mailbox details for a specified organization and user.

GetMobileDeviceAccessEffect

Simulates the effect of the mobile device access rules for the given attributes of a sample access event. Use this method to test the effects of the current set of mobile device access

ListAccessControlRules

Lists the access control rules for the specified organization.

ListAliases

Creates a paginated call to list the aliases associated with a given

ListGroupMembers

Returns an overview of the members of a group. Users and groups can be members of a

ListGroups

Returns summaries of the organization's groups.

ListMailboxExportJobs

Lists the mailbox export jobs started for the specified organization within the last

ListMailboxPermissions

Lists the mailbox permissions associated with a user, group, or resource

ListMobileDeviceAccessRules

Lists the mobile device access rules for the specified Amazon WorkMail organization.

ListOrganizations

Returns summaries of the customer's organizations.

ListResourceDelegates

Lists the delegates associated with a resource. Users and groups can be resource

ListResources

Returns summaries of the organization's resources.

ListTagsForResource

Lists the tags applied to an Amazon WorkMail organization resource.

ListUsers

Returns summaries of the organization's users.

PutAccessControlRule

Adds a new access control rule for the specified organization. The rule allows or

PutMailboxPermissions

Sets permissions for a user, group, or resource. This replaces any pre-existing

PutRetentionPolicy

Puts a retention policy to the specified organization.

RegisterToWorkMail

Registers an existing and disabled user, group, or resource for Amazon WorkMail use by

ResetPassword

Allows the administrator to reset the password for a user.

StartMailboxExportJob

Starts a mailbox export job to export MIME-format email messages and calendar items

TagResource

Applies the specified tags to the specified Amazon WorkMail organization

UntagResource

Untags the specified tags from the specified Amazon WorkMail organization

UpdateMailboxQuota

Updates a user's current mailbox quota for a specified organization and

UpdateMobileDeviceAccessRule

Updates a mobile device access rule for the specified Amazon WorkMail organization.

UpdatePrimaryEmailAddress

Updates the primary email for a user, group, or resource. The current email is moved

UpdateResource

Updates data for the resource. To have the latest information, it must be preceded by